Added hand class
