+++ test.rb
@@ -3,13 +3,13 @@ require_relative 'player'
 require_relative 'card'
 require_relative 'game'
 
-deck = CardDeck.new
-player = Player.new(name: 'Andrew')
-player.take_card(deck.take_rand_card)
-player.take_card(deck.take_ace)
-player.take_card(deck.take_ace)
-player.cards.each { |card| puts card.inspect }
-puts player.score
+# deck = CardDeck.new
+# player = Player.new(name: 'Andrew')
+# player.take_card(deck.take_rand_card)
+# player.take_card(deck.take_ace)
+# player.take_card(deck.take_ace)
+# player.cards.each { |card| puts card.inspect }
+# puts player.score
 
 # player = Player.new(name: 'user')
 # player2 = Player.new(name: 'diller')
